organize code better and add mutate function
