+++ conflict.h
@@ -1,0 +1,6 @@
+#ifndef CONFLICT_H_
+#define CONFLICT_H_
+
+int* getConflictsPerQueen(int N, int* board);
+
+#endif
