+++ test.h
@@ -1,0 +1,12 @@
+#ifndef TEST_H_
+#define TEST_H_
+
+void printBoard (int N, int* board);
+void fillRandom (int N, int* board);
+void testMutateTrulyRandom();
+void testCrossoverRandomSplit();
+void testGetConflictsPerQueen();
+void testMax();
+void testMutate();
+
+#endif
